# WIP: log values from context