Having to add a modified logger to a context when adding some values
to the context which are meant to be logged (like trace ID) has two
drawbacks:

- the instrumentation code which adds the values to the context must
   be aware of logging
- modifying the logger incurs a cost, whether some actual log entry
   then gets emitted or not.
    
A better approach is to add the values only to the context, then during logging
extract them. This is the same approach that contextual logging in Go is going
to use.

I have experimented with a few different approaches for this, including doing it entirely [in klog by wrapping the LogSink](https://github.com/kubernetes/klog/pull/358) in a logger with one that injects the extra values. That is complicated and not worth discussing further.

In logr itself there are different approaches - see individual commits in https://github.com/pohly/logr/commits/with-context-all

The one proposed here implements logging of the additional values in Logger (no need to modify LogSinks) and passes them to the LogSink via WithValues.