@pohly 2022-11-18
Here's a comparison of all benchmarks between master (b3dc695bce6838cb3f27ac700c1017947bc3235b) and this PR:

```
name                               old time/op    new time/op    delta
pkg:github.com/go-logr/logr/benchmark goos:linux goarch:amd64
DiscardLogInfoOneArg-36              95.6ns ± 4%    99.7ns ± 4%   +4.31%  (p=0.032 n=5+5)
DiscardLogInfoSeveralArgs-36          241ns ± 1%     245ns ± 1%   +1.56%  (p=0.016 n=5+5)
DiscardLogInfoWithValues-36           244ns ± 0%     249ns ± 1%   +2.35%  (p=0.008 n=5+5)
DiscardLogV0Info-36                   242ns ± 1%     252ns ± 1%   +4.20%  (p=0.008 n=5+5)
DiscardLogV9Info-36                   243ns ± 2%     263ns ± 2%   +8.03%  (p=0.008 n=5+5)
DiscardLogError-36                    270ns ± 1%     264ns ± 1%   -2.23%  (p=0.008 n=5+5)
DiscardWithValues-36                  118ns ± 1%     116ns ± 2%   -1.90%  (p=0.024 n=5+5)
DiscardWithName-36                   2.26ns ± 0%    2.29ns ± 0%   +1.43%  (p=0.008 n=5+5)
FuncrLogInfoOneArg-36                2.19µs ± 1%    2.13µs ± 1%   -3.12%  (p=0.008 n=5+5)
FuncrJSONLogInfoOneArg-36            2.47µs ± 0%    2.41µs ± 1%   -2.68%  (p=0.008 n=5+5)
FuncrLogInfoSeveralArgs-36           4.54µs ± 3%    4.52µs ± 3%     ~     (p=1.000 n=5+5)
FuncrJSONLogInfoSeveralArgs-36       4.79µs ± 3%    4.80µs ± 3%     ~     (p=1.000 n=5+5)
FuncrLogInfoWithValues-36            4.67µs ± 5%    4.59µs ± 1%     ~     (p=0.151 n=5+5)
FuncrJSONLogInfoWithValues-36        4.85µs ± 2%    4.80µs ± 4%     ~     (p=1.000 n=5+5)
FuncrLogV0Info-36                    4.56µs ± 4%    4.49µs ± 3%     ~     (p=0.421 n=5+5)
FuncrJSONLogV0Info-36                4.80µs ± 3%    4.77µs ± 3%     ~     (p=0.794 n=5+5)
FuncrLogV9Info-36                     262ns ± 1%     276ns ± 1%   +5.40%  (p=0.008 n=5+5)
FuncrJSONLogV9Info-36                 259ns ± 1%     276ns ± 1%   +6.89%  (p=0.008 n=5+5)
FuncrLogError-36                     4.89µs ± 3%    4.88µs ± 1%     ~     (p=0.841 n=5+5)
FuncrJSONLogError-36                 5.13µs ± 2%    5.11µs ± 1%     ~     (p=1.000 n=5+5)
FuncrWithValues-36                   1.40µs ± 2%    1.26µs ± 4%   -9.60%  (p=0.008 n=5+5)
FuncrWithName-36                      212ns ± 2%     171ns ± 1%  -19.41%  (p=0.008 n=5+5)
FuncrWithCallDepth-36                 210ns ± 0%     176ns ± 1%  -15.86%  (p=0.008 n=5+5)
FuncrJSONLogInfoStringerValue-36     2.64µs ± 1%    2.59µs ± 1%   -1.69%  (p=0.008 n=5+5)
FuncrJSONLogInfoErrorValue-36        2.66µs ± 2%    2.60µs ± 2%   -2.35%  (p=0.032 n=5+5)
FuncrJSONLogInfoMarshalerValue-36    2.64µs ± 2%    2.61µs ± 0%     ~     (p=0.095 n=5+5)
pkg:github.com/go-logr/logr/examples/benchmarks goos:linux goarch:amd64
NewContext1Percent-36                 181µs ± 4%      32µs ± 3%  -82.48%  (p=0.008 n=5+5)
NewContext100PercentDisabled-36       201µs ± 2%      49µs ± 3%  -75.73%  (p=0.008 n=5+5)
NewContext100Percent-36               532µs ± 2%     540µs ± 2%     ~     (p=0.095 n=5+5)

name                               old alloc/op   new alloc/op   delta
pkg:github.com/go-logr/logr/benchmark goos:linux goarch:amd64
DiscardLogInfoOneArg-36               32.0B ± 0%     32.0B ± 0%     ~     (all equal)
DiscardLogInfoSeveralArgs-36           176B ± 0%      176B ± 0%     ~     (all equal)
DiscardLogInfoWithValues-36            176B ± 0%      176B ± 0%     ~     (all equal)
DiscardLogV0Info-36                    176B ± 0%      176B ± 0%     ~     (all equal)
DiscardLogV9Info-36                    176B ± 0%      176B ± 0%     ~     (all equal)
DiscardLogError-36                     176B ± 0%      176B ± 0%     ~     (all equal)
DiscardWithValues-36                  64.0B ± 0%     64.0B ± 0%     ~     (all equal)
DiscardWithName-36                    0.00B          0.00B          ~     (all equal)
FuncrLogInfoOneArg-36                1.15kB ± 0%    1.15kB ± 0%     ~     (all equal)
FuncrJSONLogInfoOneArg-36            1.17kB ± 0%    1.17kB ± 0%     ~     (all equal)
FuncrLogInfoSeveralArgs-36           1.45kB ± 0%    1.45kB ± 0%     ~     (all equal)
FuncrJSONLogInfoSeveralArgs-36       1.45kB ± 0%    1.45kB ± 0%     ~     (all equal)
FuncrLogInfoWithValues-36            1.48kB ± 0%    1.48kB ± 0%     ~     (all equal)
FuncrJSONLogInfoWithValues-36        1.50kB ± 0%    1.50kB ± 0%     ~     (all equal)
FuncrLogV0Info-36                    1.45kB ± 0%    1.45kB ± 0%     ~     (all equal)
FuncrJSONLogV0Info-36                1.45kB ± 0%    1.45kB ± 0%     ~     (all equal)
FuncrLogV9Info-36                      176B ± 0%      176B ± 0%     ~     (all equal)
FuncrJSONLogV9Info-36                  176B ± 0%      176B ± 0%     ~     (all equal)
FuncrLogError-36                     1.48kB ± 0%    1.48kB ± 0%     ~     (all equal)
FuncrJSONLogError-36                 1.50kB ± 0%    1.50kB ± 0%     ~     (all equal)
FuncrWithValues-36                     328B ± 0%      264B ± 0%  -19.51%  (p=0.008 n=5+5)
FuncrWithName-36                       160B ± 0%       96B ± 0%  -40.00%  (p=0.008 n=5+5)
FuncrWithCallDepth-36                  160B ± 0%       96B ± 0%  -40.00%  (p=0.008 n=5+5)
FuncrJSONLogInfoStringerValue-36     1.20kB ± 0%    1.20kB ± 0%     ~     (all equal)
FuncrJSONLogInfoErrorValue-36        1.20kB ± 0%    1.20kB ± 0%     ~     (all equal)
FuncrJSONLogInfoMarshalerValue-36    1.20kB ± 0%    1.20kB ± 0%     ~     (all equal)
pkg:github.com/go-logr/logr/examples/benchmarks goos:linux goarch:amd64
NewContext1Percent-36                49.0kB ± 0%    11.2kB ± 0%  -77.14%  (p=0.008 n=5+5)
NewContext100PercentDisabled-36      57.3kB ± 0%    19.2kB ± 0%  -66.51%  (p=0.008 n=5+5)
NewContext100Percent-36               178kB ± 0%     170kB ± 0%     ~     (p=0.079 n=4+5)

name                               old allocs/op  new allocs/op  delta
pkg:github.com/go-logr/logr/benchmark goos:linux goarch:amd64
DiscardLogInfoOneArg-36                1.00 ± 0%      1.00 ± 0%     ~     (all equal)
DiscardLogInfoSeveralArgs-36           2.00 ± 0%      2.00 ± 0%     ~     (all equal)
DiscardLogInfoWithValues-36            2.00 ± 0%      2.00 ± 0%     ~     (all equal)
DiscardLogV0Info-36                    2.00 ± 0%      2.00 ± 0%     ~     (all equal)
DiscardLogV9Info-36                    2.00 ± 0%      2.00 ± 0%     ~     (all equal)
DiscardLogError-36                     2.00 ± 0%      2.00 ± 0%     ~     (all equal)
DiscardWithValues-36                   1.00 ± 0%      1.00 ± 0%     ~     (all equal)
DiscardWithName-36                     0.00           0.00          ~     (all equal)
FuncrLogInfoOneArg-36                  7.00 ± 0%      7.00 ± 0%     ~     (all equal)
FuncrJSONLogInfoOneArg-36              8.00 ± 0%      8.00 ± 0%     ~     (all equal)
FuncrLogInfoSeveralArgs-36             17.0 ± 0%      17.0 ± 0%     ~     (all equal)
FuncrJSONLogInfoSeveralArgs-36         18.0 ± 0%      18.0 ± 0%     ~     (all equal)
FuncrLogInfoWithValues-36              17.0 ± 0%      17.0 ± 0%     ~     (all equal)
FuncrJSONLogInfoWithValues-36          18.0 ± 0%      18.0 ± 0%     ~     (all equal)
FuncrLogV0Info-36                      17.0 ± 0%      17.0 ± 0%     ~     (all equal)
FuncrJSONLogV0Info-36                  18.0 ± 0%      18.0 ± 0%     ~     (all equal)
FuncrLogV9Info-36                      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
FuncrJSONLogV9Info-36                  2.00 ± 0%      2.00 ± 0%     ~     (all equal)
FuncrLogError-36                       19.0 ± 0%      19.0 ± 0%     ~     (all equal)
FuncrJSONLogError-36                   20.0 ± 0%      20.0 ± 0%     ~     (all equal)
FuncrWithValues-36                     8.00 ± 0%      8.00 ± 0%     ~     (all equal)
FuncrWithName-36                       1.00 ± 0%      1.00 ± 0%     ~     (all equal)
FuncrWithCallDepth-36                  1.00 ± 0%      1.00 ± 0%     ~     (all equal)
FuncrJSONLogInfoStringerValue-36       10.0 ± 0%      10.0 ± 0%     ~     (all equal)
FuncrJSONLogInfoErrorValue-36          10.0 ± 0%      10.0 ± 0%     ~     (all equal)
FuncrJSONLogInfoMarshalerValue-36      10.0 ± 0%      10.0 ± 0%     ~     (all equal)
pkg:github.com/go-logr/logr/examples/benchmarks goos:linux goarch:amd64
NewContext1Percent-36                 1.01k ± 0%     0.22k ± 0%  -78.24%  (p=0.008 n=5+5)
NewContext100PercentDisabled-36       1.10k ± 0%     0.30k ± 0%  -72.73%  (p=0.008 n=5+5)
NewContext100Percent-36               2.10k ± 0%     2.20k ± 0%   +4.76%  (p=0.008 n=5+5)
```